requirements.txt
fastapi
uvicorn
sqlalchemy
jinja2
python-multipart
python-dotenv
alembic